Fixed auto-rating growth mechanics in games

+++ database.js
@@ -209,7 +209,7 @@ let db = new sqlite3.Database(DBSOURCE, (err) => {
       // SR fields migration — silently ignored if columns already exist
       db.run(`ALTER TABLE entries ADD COLUMN ease_factor REAL DEFAULT 2.5`, () => {});
       db.run(`ALTER TABLE entries ADD COLUMN interval_days INTEGER DEFAULT 0`, () => {});
-      db.run(`ALTER TABLE entries ADD COLUMN repetitions INTEGER DEFAULT 0`, () => {});
+      db.run(`ALTER TABLE entries ADD COLUMN repetitions REAL DEFAULT 0`, () => {}); // REAL: weighted modes write fractional values
       db.run(`ALTER TABLE entries ADD COLUMN next_review_at TEXT DEFAULT NULL`, () => {});
       db.run(`ALTER TABLE entries ADD COLUMN last_reviewed_at TEXT DEFAULT NULL`, () => {});
